test(tui): pin above-viewport in-place rendering in the node:test suite
Move the guarding tests into tui-render.test.ts so the package test gate
(node --test) exercises them — a re-vendor that clobbered the fix would
otherwise pass acceptance — and record the behavior as divergence 6 in
the vendored-fork ledger.
